Follow-up to #81: delete Finder-dupe slop, restore research JSONs, package the scenarios pipeline, document the CLI
- Remove all 24 tracked 'foo 2.py' Finder copies (never imported or
  collected, which is how they dodged the gate); repo_layout_test now
  rejects any tracked ' N.ext' duplicate.
- Restore trace_scaling_results and the other pre-ablation research
  JSONs that the artifact sweep over-reached into; the not-committed
  policy is narrowed to .agents/docs/research/distill/ only.
- wmh/scenarios becomes proper subpackages, one per pipeline stage:
  mining/ (facets, clustering, selection), synthesis/ (scenario_set
  data model + synthesizer), verification/ (judge + verify), with
  builder.py orchestrating on top. Public API of wmh.scenarios (and
  wmh.scenarios.synthesis / .verification module paths) unchanged;
  callers of the flattened modules migrated.
- README documents 'wmh scenarios build / verify' and the model roles.
