+++ .agents/docs/research/trace_scaling_results/swe-bench.json
@@ -1,0 +1,114 @@
+{
+  "name": "trace-scaling-law",
+  "seeds": [
+    0,
+    1
+  ],
+  "conditions": [
+    {
+      "condition": {
+        "label": "base@10",
+        "params": {
+          "mode": "base",
+          "n_train": 10
+        }
+      },
+      "per_seed": [
+        {
+          "seed": 0,
+          "score": 0.7301454545454545
+        },
+        {
+          "seed": 1,
+          "score": 0.7004545454545454
+        }
+      ],
+      "mean": 0.7153,
+      "std": 0.01484545454545455
+    },
+    {
+      "condition": {
+        "label": "base@25",
+        "params": {
+          "mode": "base",
+          "n_train": 25
+        }
+      },
+      "per_seed": [
+        {
+          "seed": 0,
+          "score": 0.7405272727272727
+        },
+        {
+          "seed": 1,
+          "score": 0.7076545454545454
+        }
+      ],
+      "mean": 0.7240909090909091,
+      "std": 0.016436363636363638
+    },
+    {
+      "condition": {
+        "label": "base@50",
+        "params": {
+          "mode": "base",
+          "n_train": 50
+        }
+      },
+      "per_seed": [
+        {
+          "seed": 0,
+          "score": 0.736509090909091
+        },
+        {
+          "seed": 1,
+          "score": 0.7144363636363635
+        }
+      ],
+      "mean": 0.7254727272727273,
+      "std": 0.011036363636363733
+    },
+    {
+      "condition": {
+        "label": "base@100",
+        "params": {
+          "mode": "base",
+          "n_train": 100
+        }
+      },
+      "per_seed": [
+        {
+          "seed": 0,
+          "score": 0.7298
+        },
+        {
+          "seed": 1,
+          "score": 0.723309090909091
+        }
+      ],
+      "mean": 0.7265545454545455,
+      "std": 0.003245454545454496
+    },
+    {
+      "condition": {
+        "label": "base@170",
+        "params": {
+          "mode": "base",
+          "n_train": 170
+        }
+      },
+      "per_seed": [
+        {
+          "seed": 0,
+          "score": 0.7278363636363636
+        },
+        {
+          "seed": 1,
+          "score": 0.7319818181818182
+        }
+      ],
+      "mean": 0.7299090909090908,
+      "std": 0.0020727272727272816
+    }
+  ]
+}
