+++ .agents/docs/research/trace_scaling_results/tau-bench.json
@@ -1,0 +1,1 @@
+{"name": "trace-scaling-law", "conditions": [{"condition": {"label": "base@10"}, "mean": 0.8955, "std": 0.004500000000000004, "per_seed": [0.9, 0.891]}, {"condition": {"label": "base@25"}, "mean": 0.897, "std": 0.0010000000000000009, "per_seed": [0.896, 0.898]}, {"condition": {"label": "base@50"}, "mean": 0.896, "std": 0.0040000000000000036, "per_seed": [0.892, 0.9]}, {"condition": {"label": "base@100"}, "mean": 0.895, "std": 0.0050000000000000044, "per_seed": [0.9, 0.89]}, {"condition": {"label": "base@200"}, "mean": 0.9055, "std": 0.006500000000000006, "per_seed": [0.899, 0.912]}, {"condition": {"label": "base@400"}, "mean": 0.9155, "std": 0.0025000000000000022, "per_seed": [0.918, 0.913]}, {"condition": {"label": "base@640"}, "mean": 0.915, "std": 0.0030000000000000027, "per_seed": [0.912, 0.918]}]}
