+++ wmh/scenarios/__init__.py
@@ -1,6 +1,7 @@
 """Scenario-set construction: distill a trace corpus into a representative eval scenario set.
 
-The pipeline (Clio-style facets -> embed -> cluster -> select -> synthesize -> verify):
+The pipeline (Clio-style facets -> embed -> cluster -> select -> synthesize -> verify), organized
+as one subpackage per stage — `mining/`, `synthesis/`, `verification/` — with `builder` on top:
 
     facets = FacetExtractor(provider).extract_all(traces)
     scenario_set = build_scenario_set(traces, facets, provider, embedder, config)
@@ -10,9 +11,19 @@ Exposed via `wmh scenarios build` / `wmh scenarios verify` on the CLI.
 """
 
 from wmh.scenarios.builder import ScenarioBuildConfig, build_scenario_set
-from wmh.scenarios.clustering import TraceCluster, cluster_facets, name_clusters
-from wmh.scenarios.facets import FacetExtractor, Outcome, TraceFacet, tool_signature, trace_digest
-from wmh.scenarios.selection import SelectedTrace, hybrid_select, semdedup_keep
+from wmh.scenarios.mining import (
+    FacetExtractor,
+    Outcome,
+    SelectedTrace,
+    TraceCluster,
+    TraceFacet,
+    cluster_facets,
+    hybrid_select,
+    name_clusters,
+    semdedup_keep,
+    tool_signature,
+    trace_digest,
+)
 from wmh.scenarios.synthesis import EvalScenario, ScenarioSet, ScenarioSynthesizer
 from wmh.scenarios.verification import (
     ChecklistJudge,
