+++ wmh/scenarios/synthesis/__init__.py
@@ -1,0 +1,6 @@
+"""Synthesis: write self-contained, judgeable scenarios from selected traces."""
+
+from wmh.scenarios.synthesis.scenario_set import EvalScenario, ScenarioSet
+from wmh.scenarios.synthesis.synthesizer import ScenarioSynthesizer
+
+__all__ = ["EvalScenario", "ScenarioSet", "ScenarioSynthesizer"]
