+++ wmh/scenarios/verification/__init__.py
@@ -1,0 +1,17 @@
+"""Verification: back-agreement + solvability gates and the checklist judge that powers them."""
+
+from wmh.scenarios.verification.judge import CHECKLIST_SYSTEM, ChecklistJudge, ChecklistResult
+from wmh.scenarios.verification.verify import (
+    ScenarioVerdict,
+    VerificationReport,
+    verify_scenarios,
+)
+
+__all__ = [
+    "CHECKLIST_SYSTEM",
+    "ChecklistJudge",
+    "ChecklistResult",
+    "ScenarioVerdict",
+    "VerificationReport",
+    "verify_scenarios",
+]
