+++ wmh/scenarios/verification/verify_test.py
@@ -1,11 +1,11 @@
-"""Tests for scenario verification (checklist judge, back-agreement, solvability loop)."""
+"""Tests for the verification loop (back-agreement, solvability, report semantics)."""
 
 from __future__ import annotations
 
 from wmh.core.types import Action, ActionKind, EnvState, Observation, Step, Trace
 from wmh.engine.world_model import WorldModel
 from wmh.providers.base import Completion, Message, ProviderConfig, ProviderKind
-from wmh.scenarios.facets import Outcome
+from wmh.scenarios.mining.facets import Outcome
 from wmh.scenarios.synthesis import EvalScenario, ScenarioSet
 from wmh.scenarios.verification import (
     ChecklistJudge,
@@ -22,6 +22,7 @@ class ScriptedProvider:
         self.config = ProviderConfig(kind=ProviderKind.ANTHROPIC, model="m")
         self._replies = replies
         self.calls: list[tuple[str, str]] = []
+        self.last_max_tokens = 0
 
     def complete(
         self,
@@ -32,6 +33,7 @@ class ScriptedProvider:
         max_tokens: int = 8192,
     ) -> Completion:
         self.calls.append((system, messages[0].content))
+        self.last_max_tokens = max_tokens
         reply = self._replies[min(len(self.calls) - 1, len(self._replies) - 1)]
         return Completion(text=reply)
 
@@ -75,21 +77,6 @@ def _judge_reply(*, success: bool, passed: list[bool]) -> str:
         f'{{"passed": [{passed_json}], "success": {"true" if success else "false"}, '
         f'"critique": "graded"}}'
     )
-
-
-def test_checklist_judge_pads_short_verdicts_with_failures() -> None:
-    provider = ScriptedProvider([_judge_reply(success=True, passed=[True])])
-    result = ChecklistJudge(provider).score("task", ["a", "b", "c"], [])
-    assert result.passed == [True, False, False]
-    assert result.success is True
-    assert result.pass_rate == 1 / 3
-
-
-def test_checklist_judge_treats_garbage_as_failure() -> None:
-    result = ChecklistJudge(ScriptedProvider(["??"])).score("task", ["a"], [])
-    assert result.passed == [False]
-    assert result.success is False
-    assert "Unparseable" in result.critique
 
 
 def _scenario() -> EvalScenario:
